feat: Update PDF image extraction with PyMuPDF
The code changes include:
- Renaming `imgUtils.py` to `contentUtils.py`
- Adding a video image extraction function
- Removing the `pdfExtracter.py` file
- Adding a `createDataURL` function to convert image files to data URLs
- Refactoring the `pdf2ImageList` function to optimize extraction with offset and limit parameters
- Using the PyMuPDF library for PDF handling

These changes improve the code organization, readability, and performance of PDF image extraction.
